+++ silogistik2106752073/src/main/java/apap/ti/silogistik2106752073/dto/DetailPermintaanPengirimanDTO.java
@@ -20,7 +20,7 @@ public class DetailPermintaanPengirimanDTO {
     private String namaPenerima;
     private String alamatPenerima;
     private String jenisLayanan;
-    private Double biayaPengiriman;
+    private Integer biayaPengiriman;
     private List<BarangDetailDTO> daftarBarang;
 
 }
