+++ silogistik2106752073/src/main/java/apap/ti/silogistik2106752073/dto/ReadListBarangDTO.java
@@ -12,7 +12,7 @@ public class ReadListBarangDTO {
     private String sku;
     private String merk;
     private Integer stok;
-    private Long hargaBarang;
+    private Integer hargaBarang;
 
     // Getter and Setter
 }
