+++ silogistik2106752073/src/main/java/apap/ti/silogistik2106752073/model/PermintaanPengiriman.java
@@ -39,7 +39,7 @@ public class PermintaanPengiriman {
     private Date tanggalPengiriman;
 
     @Column(name = "biaya_pengiriman", nullable = false)
-    private Double biayaPengiriman;
+    private Integer biayaPengiriman;
 
     @Column(name = "jenis_layanan", nullable = false)
     private Integer jenisLayanan;
